ENGR00215174 implement fsl hwcomposer code.
the original hwcomposer code is a wrapper of viv hwcomposer.
now, add the real hwcomposer code.
and add the property sys.fsl.hwc to switch between the two hwcomposer.
the default hwcomposer is still a wrapper.
when the fsl hwcomposer is stable and has good performance,
it will switch to that by default.

Signed-off-by: Liu Xiaowen <b37945@freescale.com>